api cap: Block users who exceed 2000 API calls
Zoetrope is conducive to fairly high API usage.
As a security and cost-minimization measure, this
patch limits each user to approximately 2000 API calls,
after which they are notified to contact the developer
for support.

A developer (or any user with access to the Zoetrope's
Django Shell) will be able to manually reset this cap for
any user.